exceptions catched in controller

+++ server/business/src/main/java/service/TutorialService.java
@@ -10,12 +10,17 @@ import dto.mapper.TutorialMaterialMapper;
 import entity.Tutorial;
 import entity.TutorialMaterial;
 import entity.User;
+import exception.types.EntityNotFoundException;
+import exception.types.NoDataException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
+
+import static exception.Constants.*;
 
 @Service
 public class TutorialService {
@@ -33,23 +38,27 @@ public class TutorialService {
     private TutorialMaterialMapper tutorialMaterialMapper = TutorialMaterialMapper.INSTANCE;
 
     public List<TutorialDto> getAllPublicTutorials() {
+
         return tutorialMapper.entitiesToDTOs(tutorialRepository.findByIsDraft(false));
     }
 
     public List<TutorialDto> filterByKeyword(String keyword) {
+
         return tutorialMapper.entitiesToDTOs(tutorialRepository.findByKeywordsContainingIgnoreCase(keyword));
     }
 
-    public TutorialDto addTutorial(TutorialDto tutorialDto, List<String> contactPersonMsgMail) {
+    public TutorialDto addTutorial(TutorialDto tutorialDto, List<String> contactPersonMsgMail) throws EntityNotFoundException {
+
         Tutorial tutorial = tutorialMapper.mapToEntity(tutorialDto, new Tutorial());
         tutorial.setContactPersons(getUsersByMsgEmail(contactPersonMsgMail));
-        if (tutorial.getDraft()==null) {
+        if (tutorial.getDraft() == null) {
             tutorial.setDraft(false);
         }
         return tutorialMapper.mapToDTO(tutorialRepository.save(tutorial));
     }
 
     private List<User> getUsersByIds(List<Integer> contactPersonsIds) {
+
         List<User> users = new ArrayList<>();
         for (Integer id : contactPersonsIds) {
             users.add(userRepository.findOne(id));
@@ -57,52 +66,85 @@ public class TutorialService {
         return users;
     }
 
-    private List<User> getUsersByMsgEmail(List<String> contactPersonMsgMails) {
+    private List<User> getUsersByMsgEmail(List<String> contactPersonMsgMails) throws EntityNotFoundException {
+
         List<User> users = new ArrayList<>();
         for (String contactPersonMsgMail : contactPersonMsgMails) {
-            users.add(userRepository.findByMsgMail(contactPersonMsgMail).get());
+            Optional<User> optionalUser = userRepository.findByMsgMail(contactPersonMsgMail);
+            if (!optionalUser.isPresent()) {
+                throw new EntityNotFoundException(userNotFound(contactPersonMsgMail));
+            }
+            users.add(optionalUser.get());
         }
         return users;
     }
 
     public TutorialMaterialDTO addTutorialMaterial(TutorialMaterialDTO tutorialMaterialDTO) {
-        TutorialMaterial tutorialMaterial = tutorialMaterialMapper.mapToEntity(tutorialMaterialDTO, new TutorialMaterial());
+
+        TutorialMaterial tutorialMaterial = tutorialMaterialMapper.mapToEntity(tutorialMaterialDTO,
+                                                                               new TutorialMaterial());
         return tutorialMaterialMapper.mapToDTO(tutorialMaterialDAO.persistEntity(tutorialMaterial));
     }
 
-    public TutorialDto getTutorialById(Integer tutorialId) {
+    public TutorialDto getTutorialById(Integer tutorialId) throws EntityNotFoundException {
+
         Tutorial tutorialEntity = tutorialRepository.findOne(tutorialId);
+        if(tutorialEntity==null){
+            throw new EntityNotFoundException(TUTORIAL_NOT_FOUND);
+        }
         return tutorialMapper.mapToDTO(tutorialEntity);
     }
 
-    public TutorialMaterialDTO getMaterialById(Integer id) {
-        return tutorialMaterialMapper.mapToDTO(tutorialMaterialDAO.findEntity(id));
+    public TutorialMaterialDTO getMaterialById(Integer id)  {
+
+        TutorialMaterial tutorialMaterial = tutorialMaterialDAO.findEntity(id);
+        return tutorialMaterialMapper.mapToDTO(tutorialMaterial);
     }
 
-    public List<TutorialMaterialDTO> getAllMaterialsForTutorial(Integer idTutorial) {
-        List<TutorialMaterialDTO> tutorialMaterialDTOS;
+    public List<TutorialMaterialDTO> getAllMaterialsForTutorial(
+            Integer idTutorial) throws EntityNotFoundException, NoDataException {
 
-        List<TutorialMaterial> tutorialMaterials = tutorialRepository.findOne(idTutorial).getTutorialMaterials();
-        tutorialMaterialDTOS = tutorialMaterials.stream().map(tutorial -> tutorialMaterialMapper.mapToDTO(tutorial)).collect(Collectors.toList());
+        List<TutorialMaterialDTO> tutorialMaterialDTOS;
+        Tutorial tutorialEntity = tutorialRepository.findOne(idTutorial);
+        if (tutorialEntity == null) {
+            throw new EntityNotFoundException(TUTORIAL_NOT_FOUND);
+        }
+        List<TutorialMaterial> tutorialMaterials = tutorialEntity.getTutorialMaterials();
+        if (tutorialMaterials.isEmpty()) {
+            throw new NoDataException(noMaterials(tutorialEntity.getTitleTutorial()));
+        }
+        tutorialMaterialDTOS = tutorialMaterials.stream().map(
+                tutorial -> tutorialMaterialMapper.mapToDTO(tutorial)).collect(Collectors.toList());
 
         return tutorialMaterialDTOS;
     }
 
-    public List<TutorialDto> deleteTutorial(TutorialDto tutorial) {
+    public List<TutorialDto> deleteTutorial(TutorialDto tutorial) throws EntityNotFoundException {
 
         Tutorial entity = tutorialRepository.findOne(tutorial.getIdTutorial());
+        if (entity == null) {
+            throw new EntityNotFoundException(TUTORIAL_NOT_FOUND);
+        }
         tutorialRepository.delete(entity);
         return getAllPublicTutorials();
     }
 
-    public TutorialDto updateTutorial(TutorialDto tutorialDto, List<String> contactPersons) {
+    public TutorialDto updateTutorial(TutorialDto tutorialDto,
+                                      List<String> contactPersons) throws EntityNotFoundException {
+
         Tutorial tutorial = new Tutorial();
         tutorialMapper.mapToEntity(tutorialDto, tutorial);
         tutorial.setContactPersons(getUsersByMsgEmail(contactPersons));
         return tutorialMapper.mapToDTO(tutorialRepository.save(tutorial));
     }
 
-    public List<TutorialDto> allDraftTutorialsForUser(Integer idUser) {
-        return tutorialMapper.entitiesToDTOs(tutorialRepository.getAllDraftTurorialsForUser(userRepository.findOne(idUser)));
+    public List<TutorialDto> allDraftTutorialsForUser(Integer idUser) throws NoDataException {
+
+        List<TutorialDto> draftTutorials = tutorialMapper.entitiesToDTOs(
+                tutorialRepository.getAllDraftTurorialsForUser(userRepository.findOne(idUser)));
+        if (draftTutorials.isEmpty()) {
+            throw new NoDataException(NO_DRAFT);
+        } else
+            return draftTutorials;
     }
 }
